Creates and use a kapp package repository
TL;DR
-----

Deploys Spring Pet Clinic with/from a Carvel package repository

Details
-------

Prepares a kapp package repository and deploys it to the Replicated
private registry. The repository references a package that's stored
in an imgpkg bundle in the registry (the same one we used to deploy
before). Spring Pet Clinic is installed via the package.

There were a few adjustments to make to make sure that the package
loaded correctly. A package install doesn't directly reference an
image pull secret when it pulls from a repository. I had to annotate
the install with a reference to the `replicated-registry` secret,
which we still create from the license file.